+++ vite.config.ts
@@ -6,10 +6,4 @@ import { cloudflare } from "@cloudflare/vite-plugin";
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react(), cloudflare()],
-  test: {
-    environment: "miniflare",
-    miniflare: {
-      wranglerConfigPath: "wrangler.jsonc",
-    },
-  },
 })